crud cat e sub-cat de plano de contas

+++ resources/views/cat_planocontas.blade.php
@@ -3,10 +3,6 @@
 @section("conteudo")
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.mask/1.14.11/jquery.mask.min.js"></script>
-<script type="text/javascript">
-    jQuery(function ($) {
-        $("#Conta").mask("999999999999999");
-    });
 
 </script>
 
@@ -26,36 +22,42 @@
     <div class="col-md-12">
         <div class="card">
             <div class="card-header">
-                <h4 class="card-title">Plano de Contas
-                @can('insere_planocontas')
+                <h4 class="card-title">Categorias e Sub-Categorias de Plano de Contas
+                @can('insere_cat_planocontas')
+                    <button type="button" class="btn btn-success btn-rounded float-right ml-2" data-toggle="modal"
+                        data-target="#myModalCat">
+                        <i class='fas fa-plus'></i> Categorias
+                    </button>
+                    @include('modals.modal_catplanocontas')
+                @endcan
+                @can('insere_subcat_planocontas')
                     <button type="button" class="btn btn-success btn-rounded float-right" data-toggle="modal"
                         data-target="#myModal">
-                        <i class='fas fa-plus'></i> Plano de Contas
+                        <i class='fas fa-plus'></i> SubCategorias
                     </button>
+                    @include('modals.modal_subcat_planocontas')
                 @endcan
                 </h4>
 
-                @include("modals.modal_planocontas")
+               
             </div>
             <div class="form-row col-lg-12">
                 <div>
-                    <a href="{{ url("/Cadastro/planoconta") }}" class="btn btn-sm btn-info mt-3 mr-2">Todos</a>
+                    <a href="{{ url("/Cadastro/cat_planoconta") }}" class="btn btn-sm btn-info mt-3 mr-2">Todos</a>
                 </div>
                 <div class="form-group col-lg-2">
                     <select onchange="verificar(this.value)" class="form-control input-border-bottom" id="filtro"
                         name="filtro">
                         <option>Filtro de Busca</option>
                         <option value="C">Código</option>
-                        <option value="D">Cod do Pai</option>
-                        <option value="P">Tipo de Custo</option>
-                        <option value="B">Crédito/Débito</option>
+                        <option value="D">Descriçao</option>
                     </select>
                 </div>
                 <div id="name" hidden>
-                    <form action="{{ url('/PlanoContas/busca') }}" method="post">
+                    <form action="{{ url('/CatPlanoContas/busca') }}" method="post">
                         <div class=" container">
                             <div class="input-group col-lg-12 mt-2">
-                            <input type="text" class="form-control" name="criterio" placeholder="Código do Pai">
+                            <input type="text" class="form-control" name="criterio" placeholder="Descriçao">
                                 <div class="input-group-append">
                                     <button class="btn btn-success" type="submit">OK</button>
                                 </div>
@@ -66,7 +68,7 @@
                 </div>
                 
                 <div id="cod" hidden>
-                    <form action="{{ url('/PlanoContas/busca2') }}" method="post">
+                    <form action="{{ url('/CatPlanoContas/busca2') }}" method="post">
                         <div class="container">
                             <div class="input-group col-lg-8 mt-2">
                                 <input type="text" class="form-control" name="criterio" placeholder="Código">
@@ -78,48 +80,6 @@
                         {{ csrf_field() }}
                     </form>
                 </div>
-                <div id="pag_rec" hidden>
-                    <form action="{{ url('/PlanoContas/busca3') }}" method="post">
-                        <div class="container">
-                            <div class="input-group col-lg-12 mt-2 mr-2" >
-                            <select class="form-control input-border-bottom" id="criterio"
-                                name="criterio">
-                                <option>Selecione</option>
-                                <option value="CF">Custo Fixo</option>
-                                <option value="CV">Custo Variável</option>
-                                
-                            </select>
-                                <div class="input-group-append ml-2">
-                                    <button class="btn btn-success" type="submit">OK</button>
-                                </div>
-                            </div>
-                        </div>
-                        {{ csrf_field() }}
-                    </form>
-                    
-                </div>
-                <div id="conc" hidden>
-                    <form action="{{ url('/PlanoContas/busca4') }}" method="post">
-                        <div class="container">
-                        <div class="container">
-                            <div class="input-group col-lg-12 mt-2 mr-2" >
-                                <select class="form-control input-border-bottom" id="criterio"
-                                name="criterio">
-                                <option>Selecione</option>
-                                <option value="C">Crédito</option>
-                                <option value="D">Débito</option>
-                            </select>
-                                <div class="input-group-append ml-2">
-                                    <button class="btn btn-success" type="submit">OK</button>
-                                </div>
-                            </div>
-                        </div>
-                        </div>
-                        {{ csrf_field() }}
-                    </form>
-                    
-                </div>
-               
                 <div class="form-row col-lg-12">
                 @if($criterio != "")
                     <div class="card-body">
@@ -134,34 +94,35 @@
                         <thead>
                             <tr>
                                 <th class="">Cod</th>
-                                <th class="">Codigo do Pai </th>
                                 <th class="">Descrição</th>
-                                <th class="">Credito/Debito</th>
-                                <th class="">Num de Classificação</th>
-                                <th class="">Tipo de Custo</th>
+                                <th class="">Sub-Categorias</th>
                                 <th class="">Cod Adm</th>
 
                             </tr>
                         </thead>
 
                         <tbody>
-                            @foreach($planocontas as $n)
-                            @can('view_planocontas', $n)
+                            @foreach($cat_planocontas as $n)
+                            @can('view_cat_planocontas', $n)
                                 <tr>
                                     <td class=""> {{ $n->Codigo }} </td>
-                                    <td class=""> {{ $n->CodPai }} </td>
                                     <td class=""> {{ $n->Descricao }} </td>
-                                    <td class=""> {{ $n->CD=="D"?"Débito":"Crédito" }} </td>
-                                    <td class=""> {{ $n->Conta }} </td>
-                                    <td class=""> {{ $n->Tipo_Custo=="CV"?"Variável":"Fixo"  }} </td>
+                                    <td class="">
+                                        @foreach($subcat_planocontas as $s)
+                                        @if($s->cat_id == $n->Codigo)
+                                            <ul class=""> {{ $s->Descricao }}
+                                                @can('deleta_subcat_planocontas')
+                                                    <a href="javascript:deletarRegistroS('{{ $s->Codigo }}')"
+                                                    class="" style="color:red;">X</a>
+                                                @endcan
+                                             </ul>
+                                             @endif
+                                        @endforeach
+                                    </td>
                                     <td class=""> {{ $n->user_id }} </td>
                                     <td class="">
                                         <div class="btn-group" role="group">
-                                        @can('edita_planocontas')
-                                            <a href='{{ url("/PlanoContas/editar/$n->Codigo") }}'
-                                            class="btn btn-success btn-xs mr-2" style="border-radius:2px;"><i class='far fa-edit'></i></a>
-                                        @endcan
-                                        @can('deleta_planocontas')
+                                        @can('deleta_cat_planocontas')
                                             <a href="javascript:deletarRegistro('{{ $n->Codigo }}')"
                                             class="btn btn-danger btn-xs mr-2" style="border-radius:2px;"><i class='fas fa-trash-alt'></i></a>
                                         @endcan
@@ -178,7 +139,7 @@
     </div>
 
     <div class="container">
-        {{ $planocontas->links() }}
+        {{ $cat_planocontas->links() }}
     </div>
 </div>
 
@@ -234,7 +195,51 @@
         }).then((willDelete) => {
             if (willDelete) {
                 $.ajax({
-                    url: "{{ url("PlanoContas/excluir") }}" + '/' + id,
+                    url: "{{ url("CatPlanoContas/excluir") }}" + '/' + id,
+                    type: 'DELETE',
+                    data: {
+                        '_method': 'DELETE',
+                        '_token': csrf_token
+                    },
+                    success: function () {
+                        location.reload();
+                        swal({
+                            title: "Registro deletado com sucesso!",
+                            icon: "success",
+                        });
+
+                    },
+                    error: function () {
+                        swal("Erro!", "Algo de errado aconteceu!", );
+                    }
+                });
+
+            }
+        });
+    }
+
+
+    function deletarRegistroS(id) {
+        var csrf_token = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
+        swal({
+            title: "Excluir",
+            text: "Excluir do item selecionado?",
+            icon: "warning",
+            buttons: {
+                confirm: {
+                    text: 'Sim',
+                    className: 'btn btn-success'
+                },
+                cancel: {
+                    text: 'Não',
+                    visible: true,
+                    className: 'btn btn-danger'
+                }
+            }
+        }).then((willDelete) => {
+            if (willDelete) {
+                $.ajax({
+                    url: "{{ url("SubcatPlanoContas/excluir") }}" + '/' + id,
                     type: 'DELETE',
                     data: {
                         '_method': 'DELETE',
@@ -275,17 +280,6 @@
                 name.hidden = false;
                 pag.hidden = true;
                 conc.hidden = true;
-            }else if (value == "P") {
-                cod.hidden = true;
-                name.hidden = true;
-                pag.hidden = false;
-                conc.hidden = true;
-            }
-            else if (value == "B") {
-                conc.hidden = false;
-                cod.hidden = true;
-                name.hidden = true;
-                pag.hidden = true;
             }
         };
     </script>
